[Refactor] create_post.py and get_post.py file
